+++ src/features/blog-comment/index.ts
@@ -1,0 +1,2 @@
+export { BlogCommentForm } from '#/features/blog-comment/ui/BlogCommentForm'
+export { usePostBlogComment } from '#/features/blog-comment/model/use-comment'