+++ src/pages/blog/index.ts
@@ -1,0 +1,1 @@
+export { BlogPage } from '#/pages/blog/ui/BlogPage'